+++ src/commands/cmd_model.js
@@ -23,16 +23,16 @@ module.exports = {
             prompt: "Property Type and Property Name e.g. String name - exclude ';' (Press 'return' when finished)"
         });
 
-        const properties = [];
+        const PROPERTIES = [];
 
         while (property !== '') {
-            properties.push(property);
+            PROPERTIES.push(property);
             property = await vscode.window.showInputBox({
                 prompt: "Property Type and Property Name e.g. String name - exclude ';' (Press 'return' when finished)"
             });
         }
 
-        const decleration = properties.map(prop => `\t@Column(nullable: false) ${prop};\n`);
+        const decleration = PROPERTIES.map(prop => `\t@Column(nullable: false) ${prop};\n`);
         const fileContent = `
 // TODO: Update this import with your application name to import all required aqueduct imports.
 import '../application_name.dart';
@@ -43,11 +43,11 @@ class _${modelName} {
 ${decleration.join("")}
 }
         `
-        const modelFile = "/lib/models";
+        const MODEL_FILE = "/lib/models";
 
-        const fullPath = path.join(vscode.workspace.rootPath, modelFile);
-        const fileName = modelFileName.toLowerCase() + ".dart";
+        const FULL_PATH = path.join(vscode.workspace.rootPath, MODEL_FILE);
+        const FILE_NAME = modelFileName.toLowerCase() + ".dart";
         
-        await writer.writer(fullPath, fileName, fileContent);
+        await writer.writer(FULL_PATH, FILE_NAME, fileContent);
     }
 }